requirements.txt
Django==3.0.7
xmltodict==0.12.0
PyMySQL==1.0.2
Pillow==9.5.0
pycryptodome==3.17
PyJWT==1.7.1
requests==2.22.0
APScheduler==3.7.0
django-apscheduler==0.5.1
django_db_reconnect==0.0.1
jsonpath==0.82
django-cors-headers==4.1.0